Splits theme options into separate tab-files
